rename logger
